refactor: share theme toggle controls
